[democracy] extract some stuff to helpers
